// src/api.ts
export type Coin = {
    id: string;
    symbol: string;
    name: string;
    logoUrl: string;
    price: string;
    priceTimestamp: string;
    rank: string;
    dayChangePct?: string;
    weekChangePct: string;
    monthChangePct: string;
    yearChangePct: string;
    volume: string;
    high: string;
    marketCap: string;
  };